+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.0.3
+* Added: Energy sum of power from `Out to Inverter` and `Inverter to Out`
+* Added: LED display
+* Added: Units to the values
+
 ## v0.0.2
 * Added: Get automatically the grid and battery meter, if there is only one
 * Added: Select on which phase the PV Inverter is connected to